Move OpenApi setup to convention plugin
