+++ gradle/libs.versions.toml
@@ -84,8 +84,9 @@ ucrop = { group = "com.automattic", name = "ucrop", version.ref = "ucrop" }
 # Dependencies of the included build-logic
 android-gradlePlugin = { group = "com.android.tools.build", name = "gradle", version.ref = "androidGradlePlugin" }
 kotlin-gradlePlugin = { group = "org.jetbrains.kotlin", name = "kotlin-gradle-plugin", version.ref = "kotlin" }
-vanniktech-maven-publish-plugin = { group = "com.vanniktech.maven.publish", name = "com.vanniktech.maven.publish.gradle.plugin", version.ref = "mavenPublish" }
+vanniktech-maven-publishPlugin = { group = "com.vanniktech.maven.publish", name = "com.vanniktech.maven.publish.gradle.plugin", version.ref = "mavenPublish" }
 detekt-gradlePlugin = { group = "io.gitlab.arturbosch.detekt", name = "io.gitlab.arturbosch.detekt.gradle.plugin", version.ref = "detekt" }
+openapi-generator-gradlePlugin = { group = "org.openapitools", name = "openapi-generator-gradle-plugin", version.ref = "openapi" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "androidGradlePlugin" }
@@ -105,3 +106,4 @@ vanniktech-maven-publish = { id = "com.vanniktech.maven.publish", version.ref = 
 gravatar-maven-publish = { id = "gravatar.maven.publish" }
 gravatar-android-library = { id = "gravatar.android.library" }
 gravatar-android-compose = { id = "gravatar.android.compose" }
+gravatar-openapi-generator = { id = "gravatar.openapi.generator" }
